refactor: move some platform feature code under bot.
