+++ internal/remote_control/bot/card_metadata.go
@@ -2,6 +2,7 @@ package bot
 
 import (
 	"github.com/tingly-dev/tingly-box/imbot"
+	"github.com/tingly-dev/tingly-box/internal/remote_control/bot/feature"
 )
 
 // buildActionMenuMetadata builds metadata for action menu with platform-specific card rendering
@@ -13,7 +14,7 @@ func buildActionMenuMetadata(hCtx HandlerContext, tgKeyboard interface{}, card i
 
 	// For Feishu/Lark, add card_json
 	if hCtx.Platform == imbot.PlatformFeishu || hCtx.Platform == imbot.PlatformLark {
-		renderer := NewFeishuCardRenderer()
+		renderer := feature.NewFeishuCardRenderer()
 		if cardJSON, err := renderer.Render(card); err == nil {
 			metadata["card_json"] = cardJSON
 		}
