+++ internal/remote_control/bot/feature/feishu_card_renderer.go
@@ -1,4 +1,4 @@
-package bot
+package feature
 
 import (
 	"fmt"
