+++ internal/remote_control/bot/feature/telegram_dir_browser.go
@@ -1,4 +1,4 @@
-package bot
+package feature
 
 import (
 	"context"
