+++ internal/remote_control/bot/feature/telegram_keyboard.go
@@ -1,4 +1,4 @@
-package bot
+package feature
 
 import (
 	"fmt"
