+++ internal/remote_control/bot/feature/weixin_qr_client.go
@@ -1,4 +1,4 @@
-package bot
+package feature
 
 import (
 	"context"
